Trip-controller.js для отрисовки на странице
